# swagger-spring/definition/school.swagger.yaml
openapi: 3.0.3
info:
  title: Swagger School - OpenAPI 3.0
  description: |-
    This is a sample School server based on the OpenAPI 3.0 specification.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:{port}/api
    variables:
      port:
        enum:
          - '8080'
          - '8081'
        default: '8080'
...
paths:
...
  /student/all:
    get:
      tags:
        - Student
      summary: Get all students
      description: Multiple status values can be provided with comma separated strings
      operationId: allStudents
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          description: Invalid status value
      security:
        - course_auth:
            - write:courses
            - read:courses
  /student/search/{id}:
    get:
      tags:
        - Student
      summary: Search student by Id
      description: Returns a single student
      operationId: getStudentById
      parameters:
        - name: id
          in: path
          description: ID of student to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid ID supplied
        '404':
          description: Course not found
      security:
        - api_key: [ ]
        - course_auth:
            - write:pets
            - read:pets
  /student/search-by-course/{idCourse}:
    get:
      tags:
        - Student
      summary: Get all student in a Course
      description: Returns all students in a course
      operationId: getStudentByCourseIdStudent
      parameters:
        - name: idCourse
          in: path
          description: ID of course to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          description: Invalid ID supplied
        '404':
          description: Course not found
      security:
        - api_key: [ ]
        - course_auth:
            - write:pets
            - read:pets
  /student/create:
    post:
      tags:
        - Student
      summary: Add a Student
      description: Add a student
      operationId: studentCreate
      requestBody:
        description: Create a new student
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
          application/xml:
            schema:
              $ref: '#/components/schemas/Student'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
            application/xml:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - course_auth:
            - write:courses
            - read:courses
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 13
        name:
          type: string
          description: First name of student
          example: Jonh
        lastName:
          type: string
          description: last name of student
          example: Smith
        email:
          type: string
          description: Email
          example: chris.smith@local.com
        courseId:
          type: integer
          format: int64
          example: 3
      xml:
        name: student
...
  securitySchemes:
    course_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header